requirements.txt
# Enhanced Transfer Counselor Agent System Dependencies

# Core AI framework
openai-agents>=0.2.0
openai>=1.0.0

# Data handling and serialization
pyyaml>=6.0
sqlite3  # Built into Python
python-dateutil>=2.8.0

# Optional: For advanced features
# asyncio  # Built into Python (Python 3.7+)
# threading  # Built into Python
# logging  # Built into Python
# json  # Built into Python
# uuid  # Built into Python
# dataclasses  # Built into Python (Python 3.7+)
# typing  # Built into Python (Python 3.5+)
# enum  # Built into Python (Python 3.4+)
# functools  # Built into Python
# traceback  # Built into Python
# collections  # Built into Python
# pathlib  # Built into Python (Python 3.4+)
# re  # Built into Python
# time  # Built into Python
# os  # Built into Python
# sys  # Built into Python
# argparse  # Built into Python (Python 2.7+)

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0